python-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
beautifulsoup4==4.12.3
lxml==5.2.2
python-dotenv==1.0.1
